Разбиение и перевод
- Выполнено разбиение на файлы
- Логи переведены на русский
- Возращена задержка после TaskbarCreated
