+++ SolutionIconSwitcher/SolutionIconSwitcher.csproj
@@ -51,6 +51,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="ExceptionExtensions.cs" />
+    <Compile Include="Logger.cs" />
+    <Compile Include="MessageWindow.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SwitcherPackage.cs" />
   </ItemGroup>
